Network.AWS.KinesisAnalytics.AddApplicationOutput

Description

Adds an external destination to your Amazon Kinesis Analytics application.

If you want Amazon Kinesis Analytics to deliver data from an in-application stream within your application to an external destination (such as an Amazon Kinesis stream or a Firehose delivery stream), you add the relevant configuration to your application using this operation. You can configure one or more outputs for your application. Each output configuration maps an in-application stream and an external destination.

You can use one of the output configurations to deliver data from your in-application error stream to an external destination so that you can analyze the errors. For conceptual information, see Understanding Application Output (Destination) .

Note that any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

For the limits on the number of application inputs and outputs you can configure, see Limits .

This operation requires permissions to perform the kinesisanalytics:AddApplicationOutput action.

Creating a Request

addApplicationOutput

Arguments

:: Text
aaoApplicationName

-> Natural
aaoCurrentApplicationVersionId

-> Output
aaoOutput

-> AddApplicationOutput

Creates a value of AddApplicationOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aaoApplicationName - Name of the application to which you want to add the output configuration.

aaoCurrentApplicationVersionId - Version of the application to which you want add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

aaoOutput - An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, an Amazon Kinesis stream or an Amazon Kinesis Firehose delivery stream), and record the formation to use when writing to the destination.

data AddApplicationOutput

See: addApplicationOutput smart constructor.

Request Lenses

aaoApplicationName :: Lens' AddApplicationOutput Text

Name of the application to which you want to add the output configuration.

aaoCurrentApplicationVersionId :: Lens' AddApplicationOutput Natural

Version of the application to which you want add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

aaoOutput :: Lens' AddApplicationOutput Output

An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, an Amazon Kinesis stream or an Amazon Kinesis Firehose delivery stream), and record the formation to use when writing to the destination.

Destructuring the Response

addApplicationOutputResponse

Arguments

:: Int
aaorsResponseStatus

-> AddApplicationOutputResponse

Creates a value of AddApplicationOutputResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aaorsResponseStatus - -- | The response status code.

data AddApplicationOutputResponse

See: addApplicationOutputResponse smart constructor.

Response Lenses

aaorsResponseStatus :: Lens' AddApplicationOutputResponse Int

- | The response status code.