Network.AWS.KinesisAnalytics.AddApplicationReferenceDataSource

Description

Adds a reference data source to an existing application.

Amazon Kinesis Analytics reads reference data (that is, an Amazon S3 object) and creates an in-application table within your application. In the request, you provide the source (S3 bucket name and object key name), name of the in-application table to create, and the necessary mapping information that describes how data in Amazon S3 object maps to columns in the resulting in-application table.

For conceptual information, see Configuring Application Input . For the limits on data sources you can add to your application, see Limits .

This operation requires permissions to perform the kinesisanalytics:AddApplicationOutput action.

Creating a Request

addApplicationReferenceDataSource

Arguments

:: Text
aardsApplicationName

-> Natural
aardsCurrentApplicationVersionId

-> ReferenceDataSource
aardsReferenceDataSource

-> AddApplicationReferenceDataSource

Creates a value of AddApplicationReferenceDataSource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aardsApplicationName - Name of an existing application.

aardsCurrentApplicationVersionId - Version of the application for which you are adding the reference data source. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

aardsReferenceDataSource - The reference data source can be an object in your Amazon S3 bucket. Amazon Kinesis Analytics reads the object and copies the data into the in-application table that is created. You provide an S3 bucket, object key name, and the resulting in-application table that is created. You must also provide an IAM role with the necessary permissions that Amazon Kinesis Analytics can assume to read the object from your S3 bucket on your behalf.

data AddApplicationReferenceDataSource

See: addApplicationReferenceDataSource smart constructor.

Request Lenses

aardsApplicationName :: Lens' AddApplicationReferenceDataSource Text

Name of an existing application.

aardsCurrentApplicationVersionId :: Lens' AddApplicationReferenceDataSource Natural

Version of the application for which you are adding the reference data source. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

aardsReferenceDataSource :: Lens' AddApplicationReferenceDataSource ReferenceDataSource

The reference data source can be an object in your Amazon S3 bucket. Amazon Kinesis Analytics reads the object and copies the data into the in-application table that is created. You provide an S3 bucket, object key name, and the resulting in-application table that is created. You must also provide an IAM role with the necessary permissions that Amazon Kinesis Analytics can assume to read the object from your S3 bucket on your behalf.

Destructuring the Response

addApplicationReferenceDataSourceResponse

Arguments

:: Int
aardsrsResponseStatus

-> AddApplicationReferenceDataSourceResponse

Creates a value of AddApplicationReferenceDataSourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aardsrsResponseStatus - -- | The response status code.

data AddApplicationReferenceDataSourceResponse

See: addApplicationReferenceDataSourceResponse smart constructor.

Response Lenses

aardsrsResponseStatus :: Lens' AddApplicationReferenceDataSourceResponse Int

- | The response status code.