Network.AWS.KinesisAnalytics.DeleteApplicationOutput

Description

Deletes output destination configuration from your application configuration. Amazon Kinesis Analytics will no longer write data from the corresponding in-application stream to the external output destination.

This operation requires permissions to perform the kinesisanalytics:DeleteApplicationOutput action.

Creating a Request

deleteApplicationOutput

Arguments

:: Text
daoApplicationName

-> Natural
daoCurrentApplicationVersionId

-> Text
daoOutputId

-> DeleteApplicationOutput

Creates a value of DeleteApplicationOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daoApplicationName - Amazon Kinesis Analytics application name.

daoCurrentApplicationVersionId - Amazon Kinesis Analytics application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

daoOutputId - The ID of the configuration to delete. Each output configuration that is added to the application, either when the application is created or later using the AddApplicationOutput operation, has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId .

data DeleteApplicationOutput

See: deleteApplicationOutput smart constructor.

Request Lenses

daoApplicationName :: Lens' DeleteApplicationOutput Text

Amazon Kinesis Analytics application name.

daoCurrentApplicationVersionId :: Lens' DeleteApplicationOutput Natural

Amazon Kinesis Analytics application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

daoOutputId :: Lens' DeleteApplicationOutput Text

The ID of the configuration to delete. Each output configuration that is added to the application, either when the application is created or later using the AddApplicationOutput operation, has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId .

Destructuring the Response

deleteApplicationOutputResponse

Arguments

:: Int
daorsResponseStatus

-> DeleteApplicationOutputResponse

Creates a value of DeleteApplicationOutputResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daorsResponseStatus - -- | The response status code.

data DeleteApplicationOutputResponse

See: deleteApplicationOutputResponse smart constructor.

Response Lenses

daorsResponseStatus :: Lens' DeleteApplicationOutputResponse Int

- | The response status code.