Network.AWS.KinesisAnalytics.DescribeApplication

Description

Returns information about a specific Amazon Kinesis Analytics application.

If you want to retrieve a list of all applications in your account, use the ListApplications operation.

This operation requires permissions to perform the kinesisanalytics:DescribeApplication action. You can use DescribeApplication to get the current application versionId, which you need to call other operations such as Update .

Creating a Request

describeApplication

Arguments

:: Text
daApplicationName

-> DescribeApplication

Creates a value of DescribeApplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daApplicationName - Name of the application.

data DescribeApplication

See: describeApplication smart constructor.

Request Lenses

daApplicationName :: Lens' DescribeApplication Text

Name of the application.

Destructuring the Response

describeApplicationResponse

Arguments

:: Int
darsResponseStatus

-> ApplicationDetail
darsApplicationDetail

-> DescribeApplicationResponse

Creates a value of DescribeApplicationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsResponseStatus - -- | The response status code.

darsApplicationDetail - Provides a description of the application, such as the application Amazon Resource Name (ARN), status, latest version, and input and output configuration details.

data DescribeApplicationResponse

See: describeApplicationResponse smart constructor.

Response Lenses

darsResponseStatus :: Lens' DescribeApplicationResponse Int

- | The response status code.

darsApplicationDetail :: Lens' DescribeApplicationResponse ApplicationDetail

Provides a description of the application, such as the application Amazon Resource Name (ARN), status, latest version, and input and output configuration details.