Network.AWS.KinesisAnalytics.DiscoverInputSchema

Description

Infers a schema by evaluating sample records on the specified streaming source (Amazon Kinesis stream or Amazon Kinesis Firehose delivery stream). In the response, the operation returns the inferred schema and also the sample records that the operation used to infer the schema.

You can use the inferred schema when configuring a streaming source for your application. For conceptual information, see Configuring Application Input . Note that when you create an application using the Amazon Kinesis Analytics console, the console uses this operation to infer a schema and show it in the console user interface.

This operation requires permissions to perform the kinesisanalytics:DiscoverInputSchema action.

Creating a Request

discoverInputSchema

Arguments

:: Text
disResourceARN

-> Text
disRoleARN

-> InputStartingPositionConfiguration
disInputStartingPositionConfiguration

-> DiscoverInputSchema

Creates a value of DiscoverInputSchema with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

disResourceARN - Amazon Resource Name (ARN) of the streaming source.

disRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf.

disInputStartingPositionConfiguration - Point at which you want Amazon Kinesis Analytics to start reading records from the specified streaming source discovery purposes.

data DiscoverInputSchema

See: discoverInputSchema smart constructor.

Request Lenses

disResourceARN :: Lens' DiscoverInputSchema Text

Amazon Resource Name (ARN) of the streaming source.

disRoleARN :: Lens' DiscoverInputSchema Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf.

disInputStartingPositionConfiguration :: Lens' DiscoverInputSchema InputStartingPositionConfiguration

Point at which you want Amazon Kinesis Analytics to start reading records from the specified streaming source discovery purposes.

Destructuring the Response

discoverInputSchemaResponse

Arguments

:: Int
disrsResponseStatus

-> DiscoverInputSchemaResponse

Creates a value of DiscoverInputSchemaResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

disrsRawInputRecords - Raw stream data that was sampled to infer the schema.

disrsInputSchema - Schema inferred from the streaming source. It identifies the format of the data in the streaming source and how each data element maps to corresponding columns in the in-application stream that you can create.

disrsParsedInputRecords - An array of elements, where each element corresponds to a row in a stream record (a stream record can have more than one row).

disrsResponseStatus - -- | The response status code.

data DiscoverInputSchemaResponse

See: discoverInputSchemaResponse smart constructor.

Response Lenses

disrsRawInputRecords :: Lens' DiscoverInputSchemaResponse [Text]

Raw stream data that was sampled to infer the schema.

disrsInputSchema :: Lens' DiscoverInputSchemaResponse (Maybe SourceSchema)

Schema inferred from the streaming source. It identifies the format of the data in the streaming source and how each data element maps to corresponding columns in the in-application stream that you can create.

disrsParsedInputRecords :: Lens' DiscoverInputSchemaResponse [[Text]]

An array of elements, where each element corresponds to a row in a stream record (a stream record can have more than one row).

disrsResponseStatus :: Lens' DiscoverInputSchemaResponse Int

- | The response status code.