Network.AWS.KinesisAnalytics.ListApplications

Description

Returns a list of Amazon Kinesis Analytics applications in your account. For each application, the response includes the application name, Amazon Resource Name (ARN), and status. If the response returns the HasMoreApplications value as true, you can send another request by adding the ExclusiveStartApplicationName in the request body, and set the value of this to the last application name from the previous response.

If you want detailed information about a specific application, use DescribeApplication .

This operation requires permissions to perform the kinesisanalytics:ListApplications action.

Creating a Request

listApplications :: ListApplications

Creates a value of ListApplications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laLimit - Maximum number of applications to list.

laExclusiveStartApplicationName - Name of the application to start the list with. When using pagination to retrieve the list, you don't need to specify this parameter in the first request. However, in subsequent requests, you add the last application name from the previous response to get the next page of applications.

data ListApplications

See: listApplications smart constructor.

Request Lenses

laLimit :: Lens' ListApplications (Maybe Natural)

Maximum number of applications to list.

laExclusiveStartApplicationName :: Lens' ListApplications (Maybe Text)

Name of the application to start the list with. When using pagination to retrieve the list, you don't need to specify this parameter in the first request. However, in subsequent requests, you add the last application name from the previous response to get the next page of applications.

Destructuring the Response

listApplicationsResponse

Arguments

:: Int
larsResponseStatus

-> Bool
larsHasMoreApplications

-> ListApplicationsResponse

Creates a value of ListApplicationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larsResponseStatus - -- | The response status code.

larsApplicationSummaries - List of ApplicationSummary objects.

larsHasMoreApplications - Returns true if there are more applications to retrieve.

data ListApplicationsResponse

See: listApplicationsResponse smart constructor.

Response Lenses

larsResponseStatus :: Lens' ListApplicationsResponse Int

- | The response status code.

larsApplicationSummaries :: Lens' ListApplicationsResponse [ApplicationSummary]

List of ApplicationSummary objects.

larsHasMoreApplications :: Lens' ListApplicationsResponse Bool

Returns true if there are more applications to retrieve.