Network.AWS.KinesisAnalytics.StartApplication

Description

Starts the specified Amazon Kinesis Analytics application. After creating an application, you must exclusively call this operation to start your application.

After the application starts, it begins consuming the input data, processes it, and writes the output to the configured destination.

The application status must be READY for you to start an application. You can get the application status in the console or using the DescribeApplication operation.

After you start the application, you can stop the application from processing the input by calling the StopApplication operation.

This operation requires permissions to perform the kinesisanalytics:StartApplication action.

Creating a Request

startApplication

Arguments

:: Text
saApplicationName

-> StartApplication

Creates a value of StartApplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saApplicationName - Name of the application.

saInputConfigurations - Identifies the specific input, by ID, that the application starts consuming. Amazon Kinesis Analytics starts reading the streaming source associated with the input. You can also specify where in the streaming source you want Amazon Kinesis Analytics to start reading.

data StartApplication

See: startApplication smart constructor.

Request Lenses

saApplicationName :: Lens' StartApplication Text

Name of the application.

saInputConfigurations :: Lens' StartApplication [InputConfiguration]

Identifies the specific input, by ID, that the application starts consuming. Amazon Kinesis Analytics starts reading the streaming source associated with the input. You can also specify where in the streaming source you want Amazon Kinesis Analytics to start reading.

Destructuring the Response

startApplicationResponse

Arguments

:: Int
sarsResponseStatus

-> StartApplicationResponse

Creates a value of StartApplicationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sarsResponseStatus - -- | The response status code.

data StartApplicationResponse

See: startApplicationResponse smart constructor.

Response Lenses

sarsResponseStatus :: Lens' StartApplicationResponse Int

- | The response status code.