Network.AWS.KinesisAnalytics.Types

Description

Service Configuration

kinesisAnalytics :: Service

API version 2015-08-14 of the Amazon Kinesis Analytics SDK configuration.

Errors

_InvalidApplicationConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

User-provided application configuration is not valid.

_ResourceProvisionedThroughputExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Discovery failed to get a record from the streaming source because of the Amazon Kinesis Streams ProvisionedThroughputExceededException. For more information, see GetRecords in the Amazon Kinesis Streams API Reference.

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

Specified input parameter value is invalid.

_CodeValidationException :: AsError a => Getting (First ServiceError) a ServiceError

User-provided application code (query) is invalid. This can be a simple syntax error.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Exception thrown as a result of concurrent modification to an application. For example, two individuals attempting to edit the same application at the same time.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is unavailable, back off and retry the operation.

_UnableToDetectSchemaException :: AsError a => Getting (First ServiceError) a ServiceError

Data format is not valid, Amazon Kinesis Analytics is not able to detect schema for the given streaming source.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Specified application can't be found.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exceeded the number of applications allowed.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

Application is not available for this operation.

ApplicationStatus

data ApplicationStatus

Constructors

Deleting

Ready

Running

Starting

Stopping

Updating

InputStartingPosition

data InputStartingPosition

Constructors

LastStoppedPoint

Now

TrimHorizon

RecordFormatType

data RecordFormatType

Constructors

CSV

JSON

ApplicationDetail

data ApplicationDetail

Provides a description of the application, including the application Amazon Resource Name (ARN), status, latest version, and input and output configuration.

See: applicationDetail smart constructor.

applicationDetail

Arguments

:: Text
adApplicationName

-> Text
adApplicationARN

-> ApplicationStatus
adApplicationStatus

-> Natural
adApplicationVersionId

-> ApplicationDetail

Creates a value of ApplicationDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adApplicationDescription - Description of the application.

adOutputDescriptions - Describes the application output configuration. For more information, see Configuring Application Output .

adCloudWatchLoggingOptionDescriptions - Describes the CloudWatch log streams that are configured to receive application messages. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs .

adReferenceDataSourceDescriptions - Describes reference data sources configured for the application. For more information, see Configuring Application Input .

adInputDescriptions - Describes the application input configuration. For more information, see Configuring Application Input .

adApplicationCode - Returns the application code that you provided to perform data analysis on any of the in-application streams in your application.

adCreateTimestamp - Timestamp when the application version was created.

adLastUpdateTimestamp - Timestamp when the application was last updated.

adApplicationName - Name of the application.

adApplicationARN - ARN of the application.

adApplicationStatus - Status of the application.

adApplicationVersionId - Provides the current application version.

adApplicationDescription :: Lens' ApplicationDetail (Maybe Text)

Description of the application.

adOutputDescriptions :: Lens' ApplicationDetail [OutputDescription]

Describes the application output configuration. For more information, see Configuring Application Output .

adCloudWatchLoggingOptionDescriptions :: Lens' ApplicationDetail [CloudWatchLoggingOptionDescription]

Describes the CloudWatch log streams that are configured to receive application messages. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs .

adReferenceDataSourceDescriptions :: Lens' ApplicationDetail [ReferenceDataSourceDescription]

Describes reference data sources configured for the application. For more information, see Configuring Application Input .

adInputDescriptions :: Lens' ApplicationDetail [InputDescription]

Describes the application input configuration. For more information, see Configuring Application Input .

adApplicationCode :: Lens' ApplicationDetail (Maybe Text)

Returns the application code that you provided to perform data analysis on any of the in-application streams in your application.

adCreateTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

Timestamp when the application version was created.

adLastUpdateTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

Timestamp when the application was last updated.

adApplicationName :: Lens' ApplicationDetail Text

Name of the application.

adApplicationARN :: Lens' ApplicationDetail Text

ARN of the application.

adApplicationStatus :: Lens' ApplicationDetail ApplicationStatus

Status of the application.

adApplicationVersionId :: Lens' ApplicationDetail Natural

Provides the current application version.

ApplicationSummary

data ApplicationSummary

Provides application summary information, including the application Amazon Resource Name (ARN), name, and status.

See: applicationSummary smart constructor.

applicationSummary

Arguments

:: Text
asApplicationName

-> Text
asApplicationARN

-> ApplicationStatus
asApplicationStatus

-> ApplicationSummary

Creates a value of ApplicationSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asApplicationName - Name of the application.

asApplicationARN - ARN of the application.

asApplicationStatus - Status of the application.

asApplicationName :: Lens' ApplicationSummary Text

Name of the application.

asApplicationARN :: Lens' ApplicationSummary Text

ARN of the application.

asApplicationStatus :: Lens' ApplicationSummary ApplicationStatus

Status of the application.

ApplicationUpdate

data ApplicationUpdate

Describes updates to apply to an existing Amazon Kinesis Analytics application.

See: applicationUpdate smart constructor.

applicationUpdate :: ApplicationUpdate

Creates a value of ApplicationUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

auReferenceDataSourceUpdates - Describes application reference data source updates.

auInputUpdates - Describes application input configuration updates.

auCloudWatchLoggingOptionUpdates - Describes application CloudWatch logging option updates.

auOutputUpdates - Describes application output configuration updates.

auApplicationCodeUpdate - Describes application code updates.

auReferenceDataSourceUpdates :: Lens' ApplicationUpdate [ReferenceDataSourceUpdate]

Describes application reference data source updates.

auInputUpdates :: Lens' ApplicationUpdate [InputUpdate]

Describes application input configuration updates.

auCloudWatchLoggingOptionUpdates :: Lens' ApplicationUpdate [CloudWatchLoggingOptionUpdate]

Describes application CloudWatch logging option updates.

auOutputUpdates :: Lens' ApplicationUpdate [OutputUpdate]

Describes application output configuration updates.

auApplicationCodeUpdate :: Lens' ApplicationUpdate (Maybe Text)

Describes application code updates.

CSVMappingParameters

data CSVMappingParameters

Provides additional mapping information when the record format uses delimiters, such as CSV. For example, the following sample records use CSV format, where the records use the '\n' as the row delimiter and a comma (",") as the column delimiter:

"name1", "address1"

"name2, "address2"

See: csvMappingParameters smart constructor.

csvMappingParameters

Arguments

:: Text
cmpRecordRowDelimiter

-> Text
cmpRecordColumnDelimiter

-> CSVMappingParameters

Creates a value of CSVMappingParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmpRecordRowDelimiter - Row delimiter. For example, in a CSV format, '\n' is the typical row delimiter.

cmpRecordColumnDelimiter - Column delimiter. For example, in a CSV format, a comma (",") is the typical column delimiter.

cmpRecordRowDelimiter :: Lens' CSVMappingParameters Text

Row delimiter. For example, in a CSV format, '\n' is the typical row delimiter.

cmpRecordColumnDelimiter :: Lens' CSVMappingParameters Text

Column delimiter. For example, in a CSV format, a comma (",") is the typical column delimiter.

CloudWatchLoggingOption

data CloudWatchLoggingOption

Provides a description of CloudWatch logging options, including the log stream Amazon Resource Name (ARN) and the role ARN.

See: cloudWatchLoggingOption smart constructor.

cloudWatchLoggingOption

Arguments

:: Text
cwloLogStreamARN

-> Text
cwloRoleARN

-> CloudWatchLoggingOption

Creates a value of CloudWatchLoggingOption with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cwloLogStreamARN - ARN of the CloudWatch log to receive application messages.

cwloRoleARN - IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role that is used must have the PutLogEvents policy action enabled.

cwloLogStreamARN :: Lens' CloudWatchLoggingOption Text

ARN of the CloudWatch log to receive application messages.

cwloRoleARN :: Lens' CloudWatchLoggingOption Text

IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role that is used must have the PutLogEvents policy action enabled.

CloudWatchLoggingOptionDescription

data CloudWatchLoggingOptionDescription

Description of the CloudWatch logging option.

See: cloudWatchLoggingOptionDescription smart constructor.

cloudWatchLoggingOptionDescription

Arguments

:: Text
cwlodLogStreamARN

-> Text
cwlodRoleARN

-> CloudWatchLoggingOptionDescription

Creates a value of CloudWatchLoggingOptionDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cwlodCloudWatchLoggingOptionId - ID of the CloudWatch logging option description.

cwlodLogStreamARN - ARN of the CloudWatch log to receive application messages.

cwlodRoleARN - IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role used must have the PutLogEvents policy action enabled.

cwlodCloudWatchLoggingOptionId :: Lens' CloudWatchLoggingOptionDescription (Maybe Text)

ID of the CloudWatch logging option description.

cwlodLogStreamARN :: Lens' CloudWatchLoggingOptionDescription Text

ARN of the CloudWatch log to receive application messages.

cwlodRoleARN :: Lens' CloudWatchLoggingOptionDescription Text

IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role used must have the PutLogEvents policy action enabled.

CloudWatchLoggingOptionUpdate

data CloudWatchLoggingOptionUpdate

Describes CloudWatch logging option updates.

See: cloudWatchLoggingOptionUpdate smart constructor.

cloudWatchLoggingOptionUpdate

Arguments

:: Text
cwlouCloudWatchLoggingOptionId

-> CloudWatchLoggingOptionUpdate

Creates a value of CloudWatchLoggingOptionUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cwlouRoleARNUpdate - IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role used must have the PutLogEvents policy action enabled.

cwlouLogStreamARNUpdate - ARN of the CloudWatch log to receive application messages.

cwlouCloudWatchLoggingOptionId - ID of the CloudWatch logging option to update

cwlouRoleARNUpdate :: Lens' CloudWatchLoggingOptionUpdate (Maybe Text)

IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role used must have the PutLogEvents policy action enabled.

cwlouLogStreamARNUpdate :: Lens' CloudWatchLoggingOptionUpdate (Maybe Text)

ARN of the CloudWatch log to receive application messages.

cwlouCloudWatchLoggingOptionId :: Lens' CloudWatchLoggingOptionUpdate Text

ID of the CloudWatch logging option to update

DestinationSchema

data DestinationSchema

Describes the data format when records are written to the destination. For more information, see Configuring Application Output .

See: destinationSchema smart constructor.

destinationSchema :: DestinationSchema

Creates a value of DestinationSchema with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsRecordFormatType - Specifies the format of the records on the output stream.

dsRecordFormatType :: Lens' DestinationSchema (Maybe RecordFormatType)

Specifies the format of the records on the output stream.

Input

data Input

When you configure the application input, you specify the streaming source, the in-application stream name that is created, and the mapping between the two. For more information, see Configuring Application Input .

See: input smart constructor.

input

Arguments

:: Text
iNamePrefix

-> SourceSchema
iInputSchema

-> Input

Creates a value of Input with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iInputParallelism - Describes the number of in-application streams to create. Data from your source will be routed to these in-application input streams. (see Configuring Application Input .

iKinesisStreamsInput - If the streaming source is an Amazon Kinesis stream, identifies the stream's Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf. Note: Either KinesisStreamsInput or KinesisFirehoseInput is required.

iKinesisFirehoseInput - If the streaming source is an Amazon Kinesis Firehose delivery stream, identifies the Firehose delivery stream's ARN and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf. Note: Either KinesisStreamsInput or KinesisFirehoseInput is required.

iNamePrefix - Name prefix to use when creating in-application stream. Suppose you specify a prefix MyInApplicationStream. Amazon Kinesis Analytics will then create one or more (as per the InputParallelism count you specified) in-application streams with names MyInApplicationStream_001, MyInApplicationStream_002 and so on.

iInputSchema - Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created. Also used to describe the format of the reference data source.

iInputParallelism :: Lens' Input (Maybe InputParallelism)

Describes the number of in-application streams to create. Data from your source will be routed to these in-application input streams. (see Configuring Application Input .

iKinesisStreamsInput :: Lens' Input (Maybe KinesisStreamsInput)

If the streaming source is an Amazon Kinesis stream, identifies the stream's Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf. Note: Either KinesisStreamsInput or KinesisFirehoseInput is required.

iKinesisFirehoseInput :: Lens' Input (Maybe KinesisFirehoseInput)

If the streaming source is an Amazon Kinesis Firehose delivery stream, identifies the Firehose delivery stream's ARN and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf. Note: Either KinesisStreamsInput or KinesisFirehoseInput is required.

iNamePrefix :: Lens' Input Text

Name prefix to use when creating in-application stream. Suppose you specify a prefix MyInApplicationStream. Amazon Kinesis Analytics will then create one or more (as per the InputParallelism count you specified) in-application streams with names MyInApplicationStream_001, MyInApplicationStream_002 and so on.

iInputSchema :: Lens' Input SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created. Also used to describe the format of the reference data source.

InputConfiguration

data InputConfiguration

When you start your application, you provide this configuration, which identifies the input source and the point in the input source at which you want the application to start processing records.

See: inputConfiguration smart constructor.

inputConfiguration

Arguments

:: Text
icId

-> InputStartingPositionConfiguration
icInputStartingPositionConfiguration

-> InputConfiguration

Creates a value of InputConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

icId - Input source ID. You can get this ID by calling the DescribeApplication operation.

icInputStartingPositionConfiguration - Point at which you want the application to start processing records from the streaming source.

icId :: Lens' InputConfiguration Text

Input source ID. You can get this ID by calling the DescribeApplication operation.

icInputStartingPositionConfiguration :: Lens' InputConfiguration InputStartingPositionConfiguration

Point at which you want the application to start processing records from the streaming source.

InputDescription

data InputDescription

Describes the application input configuration. For more information, see Configuring Application Input .

See: inputDescription smart constructor.

inputDescription :: InputDescription

Creates a value of InputDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idInputStartingPositionConfiguration - Point at which the application is configured to read from the input stream.

idInputParallelism - Describes the configured parallelism (number of in-application streams mapped to the streaming source).

idInputId - Input ID associated with the application input. This is the ID that Amazon Kinesis Analytics assigns to each input configuration you add to your application.

idInAppStreamNames - Returns the in-application stream names that are mapped to the stream source.

idKinesisFirehoseInputDescription - If an Amazon Kinesis Firehose delivery stream is configured as a streaming source, provides the Firehose delivery stream's Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

idInputSchema - Undocumented member.

idKinesisStreamsInputDescription - If an Amazon Kinesis stream is configured as streaming source, provides Amazon Kinesis stream's ARN and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

idNamePrefix - In-application name prefix.

idInputStartingPositionConfiguration :: Lens' InputDescription (Maybe InputStartingPositionConfiguration)

Point at which the application is configured to read from the input stream.

idInputParallelism :: Lens' InputDescription (Maybe InputParallelism)

Describes the configured parallelism (number of in-application streams mapped to the streaming source).

idInputId :: Lens' InputDescription (Maybe Text)

Input ID associated with the application input. This is the ID that Amazon Kinesis Analytics assigns to each input configuration you add to your application.

idInAppStreamNames :: Lens' InputDescription [Text]

Returns the in-application stream names that are mapped to the stream source.

idKinesisFirehoseInputDescription :: Lens' InputDescription (Maybe KinesisFirehoseInputDescription)

If an Amazon Kinesis Firehose delivery stream is configured as a streaming source, provides the Firehose delivery stream's Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

idInputSchema :: Lens' InputDescription (Maybe SourceSchema)

Undocumented member.

idKinesisStreamsInputDescription :: Lens' InputDescription (Maybe KinesisStreamsInputDescription)

If an Amazon Kinesis stream is configured as streaming source, provides Amazon Kinesis stream's ARN and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

idNamePrefix :: Lens' InputDescription (Maybe Text)

In-application name prefix.

InputParallelism

data InputParallelism

Describes the number of in-application streams to create for a given streaming source. For information about parallelism, see Configuring Application Input .

See: inputParallelism smart constructor.

inputParallelism :: InputParallelism

Creates a value of InputParallelism with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ipCount - Number of in-application streams to create. For more information, see Limits .

ipCount :: Lens' InputParallelism (Maybe Natural)

Number of in-application streams to create. For more information, see Limits .

InputParallelismUpdate

data InputParallelismUpdate

Provides updates to the parallelism count.

See: inputParallelismUpdate smart constructor.

inputParallelismUpdate :: InputParallelismUpdate

Creates a value of InputParallelismUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ipuCountUpdate - Number of in-application streams to create for the specified streaming source.

ipuCountUpdate :: Lens' InputParallelismUpdate (Maybe Natural)

Number of in-application streams to create for the specified streaming source.

InputSchemaUpdate

data InputSchemaUpdate

Describes updates for the application's input schema.

See: inputSchemaUpdate smart constructor.

inputSchemaUpdate :: InputSchemaUpdate

Creates a value of InputSchemaUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

isuRecordFormatUpdate - Specifies the format of the records on the streaming source.

isuRecordEncodingUpdate - Specifies the encoding of the records in the streaming source. For example, UTF-8.

isuRecordColumnUpdates - A list of RecordColumn objects. Each object describes the mapping of the streaming source element to the corresponding column in the in-application stream.

isuRecordFormatUpdate :: Lens' InputSchemaUpdate (Maybe RecordFormat)

Specifies the format of the records on the streaming source.

isuRecordEncodingUpdate :: Lens' InputSchemaUpdate (Maybe Text)

Specifies the encoding of the records in the streaming source. For example, UTF-8.

isuRecordColumnUpdates :: Lens' InputSchemaUpdate (Maybe (NonEmpty RecordColumn))

A list of RecordColumn objects. Each object describes the mapping of the streaming source element to the corresponding column in the in-application stream.

InputStartingPositionConfiguration

data InputStartingPositionConfiguration

Describes the point at which the application reads from the streaming source.

See: inputStartingPositionConfiguration smart constructor.

inputStartingPositionConfiguration :: InputStartingPositionConfiguration

Creates a value of InputStartingPositionConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ispcInputStartingPosition - The starting position on the stream. * NOW - Start reading just after the most recent record in the stream, start at the request timestamp that the customer issued. * TRIM_HORIZON - Start reading at the last untrimmed record in the stream, which is the oldest record available in the stream. This option is not available for an Amazon Kinesis Firehose delivery stream. * LAST_STOPPED_POINT - Resume reading from where the application last stopped reading.

ispcInputStartingPosition :: Lens' InputStartingPositionConfiguration (Maybe InputStartingPosition)

The starting position on the stream. * NOW - Start reading just after the most recent record in the stream, start at the request timestamp that the customer issued. * TRIM_HORIZON - Start reading at the last untrimmed record in the stream, which is the oldest record available in the stream. This option is not available for an Amazon Kinesis Firehose delivery stream. * LAST_STOPPED_POINT - Resume reading from where the application last stopped reading.

InputUpdate

data InputUpdate

Describes updates to a specific input configuration (identified by the InputId of an application).

See: inputUpdate smart constructor.

inputUpdate

Arguments

:: Text
iuInputId

-> InputUpdate

Creates a value of InputUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iuKinesisStreamsInputUpdate - If a Amazon Kinesis stream is the streaming source to be updated, provides an updated stream ARN and IAM role ARN.

iuInputParallelismUpdate - Describes the parallelism updates (the number in-application streams Amazon Kinesis Analytics creates for the specific streaming source).

iuNamePrefixUpdate - Name prefix for in-application streams that Amazon Kinesis Analytics creates for the specific streaming source.

iuInputSchemaUpdate - Describes the data format on the streaming source, and how record elements on the streaming source map to columns of the in-application stream that is created.

iuKinesisFirehoseInputUpdate - If an Amazon Kinesis Firehose delivery stream is the streaming source to be updated, provides an updated stream Amazon Resource Name (ARN) and IAM role ARN.

iuInputId - Input ID of the application input to be updated.

iuKinesisStreamsInputUpdate :: Lens' InputUpdate (Maybe KinesisStreamsInputUpdate)

If a Amazon Kinesis stream is the streaming source to be updated, provides an updated stream ARN and IAM role ARN.

iuInputParallelismUpdate :: Lens' InputUpdate (Maybe InputParallelismUpdate)

Describes the parallelism updates (the number in-application streams Amazon Kinesis Analytics creates for the specific streaming source).

iuNamePrefixUpdate :: Lens' InputUpdate (Maybe Text)

Name prefix for in-application streams that Amazon Kinesis Analytics creates for the specific streaming source.

iuInputSchemaUpdate :: Lens' InputUpdate (Maybe InputSchemaUpdate)

Describes the data format on the streaming source, and how record elements on the streaming source map to columns of the in-application stream that is created.

iuKinesisFirehoseInputUpdate :: Lens' InputUpdate (Maybe KinesisFirehoseInputUpdate)

If an Amazon Kinesis Firehose delivery stream is the streaming source to be updated, provides an updated stream Amazon Resource Name (ARN) and IAM role ARN.

iuInputId :: Lens' InputUpdate Text

Input ID of the application input to be updated.

JSONMappingParameters

data JSONMappingParameters

Provides additional mapping information when JSON is the record format on the streaming source.

See: jsonMappingParameters smart constructor.

jsonMappingParameters

Arguments

:: Text
jmpRecordRowPath

-> JSONMappingParameters

Creates a value of JSONMappingParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jmpRecordRowPath - Path to the top-level parent that contains the records.

jmpRecordRowPath :: Lens' JSONMappingParameters Text

Path to the top-level parent that contains the records.

KinesisFirehoseInput

data KinesisFirehoseInput

Identifies an Amazon Kinesis Firehose delivery stream as the streaming source. You provide the Firehose delivery stream's Amazon Resource Name (ARN) and an IAM role ARN that enables Amazon Kinesis Analytics to access the stream on your behalf.

See: kinesisFirehoseInput smart constructor.

kinesisFirehoseInput

Arguments

:: Text
kfiResourceARN

-> Text
kfiRoleARN

-> KinesisFirehoseInput

Creates a value of KinesisFirehoseInput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kfiResourceARN - ARN of the input Firehose delivery stream.

kfiRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to make sure the role has necessary permissions to access the stream.

kfiResourceARN :: Lens' KinesisFirehoseInput Text

ARN of the input Firehose delivery stream.

kfiRoleARN :: Lens' KinesisFirehoseInput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to make sure the role has necessary permissions to access the stream.

KinesisFirehoseInputDescription

data KinesisFirehoseInputDescription

Describes the Amazon Kinesis Firehose delivery stream that is configured as the streaming source in the application input configuration.

See: kinesisFirehoseInputDescription smart constructor.

kinesisFirehoseInputDescription :: KinesisFirehoseInputDescription

Creates a value of KinesisFirehoseInputDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kfidResourceARN - Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream.

kfidRoleARN - ARN of the IAM role that Amazon Kinesis Analytics assumes to access the stream.

kfidResourceARN :: Lens' KinesisFirehoseInputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream.

kfidRoleARN :: Lens' KinesisFirehoseInputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics assumes to access the stream.

KinesisFirehoseInputUpdate

data KinesisFirehoseInputUpdate

When updating application input configuration, provides information about an Amazon Kinesis Firehose delivery stream as the streaming source.

See: kinesisFirehoseInputUpdate smart constructor.

kinesisFirehoseInputUpdate :: KinesisFirehoseInputUpdate

Creates a value of KinesisFirehoseInputUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kfiuRoleARNUpdate - Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant necessary permissions to this role.

kfiuResourceARNUpdate - ARN of the input Amazon Kinesis Firehose delivery stream to read.

kfiuRoleARNUpdate :: Lens' KinesisFirehoseInputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant necessary permissions to this role.

kfiuResourceARNUpdate :: Lens' KinesisFirehoseInputUpdate (Maybe Text)

ARN of the input Amazon Kinesis Firehose delivery stream to read.

KinesisFirehoseOutput

data KinesisFirehoseOutput

When configuring application output, identifies an Amazon Kinesis Firehose delivery stream as the destination. You provide the stream Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to write to the stream on your behalf.

See: kinesisFirehoseOutput smart constructor.

kinesisFirehoseOutput

Arguments

:: Text
kfoResourceARN

-> Text
kfoRoleARN

-> KinesisFirehoseOutput

Creates a value of KinesisFirehoseOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kfoResourceARN - ARN of the destination Amazon Kinesis Firehose delivery stream to write to.

kfoRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

kfoResourceARN :: Lens' KinesisFirehoseOutput Text

ARN of the destination Amazon Kinesis Firehose delivery stream to write to.

kfoRoleARN :: Lens' KinesisFirehoseOutput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

KinesisFirehoseOutputDescription

data KinesisFirehoseOutputDescription

For an application output, describes the Amazon Kinesis Firehose delivery stream configured as its destination.

See: kinesisFirehoseOutputDescription smart constructor.

kinesisFirehoseOutputDescription :: KinesisFirehoseOutputDescription

Creates a value of KinesisFirehoseOutputDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kfodResourceARN - Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream.

kfodRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

kfodResourceARN :: Lens' KinesisFirehoseOutputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream.

kfodRoleARN :: Lens' KinesisFirehoseOutputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

KinesisFirehoseOutputUpdate

data KinesisFirehoseOutputUpdate

When updating an output configuration using the UpdateApplication operation, provides information about an Amazon Kinesis Firehose delivery stream configured as the destination.

See: kinesisFirehoseOutputUpdate smart constructor.

kinesisFirehoseOutputUpdate :: KinesisFirehoseOutputUpdate

Creates a value of KinesisFirehoseOutputUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kfouRoleARNUpdate - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant necessary permissions to this role.

kfouResourceARNUpdate - Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream to write to.

kfouRoleARNUpdate :: Lens' KinesisFirehoseOutputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant necessary permissions to this role.

kfouResourceARNUpdate :: Lens' KinesisFirehoseOutputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream to write to.

KinesisStreamsInput

data KinesisStreamsInput

Identifies an Amazon Kinesis stream as the streaming source. You provide the stream's ARN and an IAM role ARN that enables Amazon Kinesis Analytics to access the stream on your behalf.

See: kinesisStreamsInput smart constructor.

kinesisStreamsInput

Arguments

:: Text
ksiResourceARN

-> Text
ksiRoleARN

-> KinesisStreamsInput

Creates a value of KinesisStreamsInput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ksiResourceARN - ARN of the input Amazon Kinesis stream to read.

ksiRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

ksiResourceARN :: Lens' KinesisStreamsInput Text

ARN of the input Amazon Kinesis stream to read.

ksiRoleARN :: Lens' KinesisStreamsInput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

KinesisStreamsInputDescription

data KinesisStreamsInputDescription

Describes the Amazon Kinesis stream that is configured as the streaming source in the application input configuration.

See: kinesisStreamsInputDescription smart constructor.

kinesisStreamsInputDescription :: KinesisStreamsInputDescription

Creates a value of KinesisStreamsInputDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ksidResourceARN - Amazon Resource Name (ARN) of the Amazon Kinesis stream.

ksidRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

ksidResourceARN :: Lens' KinesisStreamsInputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis stream.

ksidRoleARN :: Lens' KinesisStreamsInputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

KinesisStreamsInputUpdate

data KinesisStreamsInputUpdate

When updating application input configuration, provides information about an Amazon Kinesis stream as the streaming source.

See: kinesisStreamsInputUpdate smart constructor.

kinesisStreamsInputUpdate :: KinesisStreamsInputUpdate

Creates a value of KinesisStreamsInputUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ksiuRoleARNUpdate - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

ksiuResourceARNUpdate - Amazon Resource Name (ARN) of the input Amazon Kinesis stream to read.

ksiuRoleARNUpdate :: Lens' KinesisStreamsInputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

ksiuResourceARNUpdate :: Lens' KinesisStreamsInputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the input Amazon Kinesis stream to read.

KinesisStreamsOutput

data KinesisStreamsOutput

When configuring application output, identifies a Amazon Kinesis stream as the destination. You provide the stream Amazon Resource Name (ARN) and also an IAM role ARN that Amazon Kinesis Analytics can use to write to the stream on your behalf.

See: kinesisStreamsOutput smart constructor.

kinesisStreamsOutput

Arguments

:: Text
ksoResourceARN

-> Text
ksoRoleARN

-> KinesisStreamsOutput

Creates a value of KinesisStreamsOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ksoResourceARN - ARN of the destination Amazon Kinesis stream to write to.

ksoRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

ksoResourceARN :: Lens' KinesisStreamsOutput Text

ARN of the destination Amazon Kinesis stream to write to.

ksoRoleARN :: Lens' KinesisStreamsOutput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

KinesisStreamsOutputDescription

data KinesisStreamsOutputDescription

For an application output, describes the Amazon Kinesis stream configured as its destination.

See: kinesisStreamsOutputDescription smart constructor.

kinesisStreamsOutputDescription :: KinesisStreamsOutputDescription

Creates a value of KinesisStreamsOutputDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ksodResourceARN - Amazon Resource Name (ARN) of the Amazon Kinesis stream.

ksodRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

ksodResourceARN :: Lens' KinesisStreamsOutputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis stream.

ksodRoleARN :: Lens' KinesisStreamsOutputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

KinesisStreamsOutputUpdate

data KinesisStreamsOutputUpdate

When updating an output configuration using the UpdateApplication operation, provides information about an Amazon Kinesis stream configured as the destination.

See: kinesisStreamsOutputUpdate smart constructor.

kinesisStreamsOutputUpdate :: KinesisStreamsOutputUpdate

Creates a value of KinesisStreamsOutputUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ksouRoleARNUpdate - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

ksouResourceARNUpdate - Amazon Resource Name (ARN) of the Amazon Kinesis stream where you want to write the output.

ksouRoleARNUpdate :: Lens' KinesisStreamsOutputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

ksouResourceARNUpdate :: Lens' KinesisStreamsOutputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis stream where you want to write the output.

MappingParameters

data MappingParameters

When configuring application input at the time of creating or updating an application, provides additional mapping information specific to the record format (such as JSON, CSV, or record fields delimited by some delimiter) on the streaming source.

See: mappingParameters smart constructor.

mappingParameters :: MappingParameters

Creates a value of MappingParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mpCSVMappingParameters - Provides additional mapping information when the record format uses delimiters (for example, CSV).

mpJSONMappingParameters - Provides additional mapping information when JSON is the record format on the streaming source.

mpCSVMappingParameters :: Lens' MappingParameters (Maybe CSVMappingParameters)

Provides additional mapping information when the record format uses delimiters (for example, CSV).

mpJSONMappingParameters :: Lens' MappingParameters (Maybe JSONMappingParameters)

Provides additional mapping information when JSON is the record format on the streaming source.

Output

data Output

Describes application output configuration in which you identify an in-application stream and a destination where you want the in-application stream data to be written. The destination can be an Amazon Kinesis stream or an Amazon Kinesis Firehose delivery stream.

For limits on how many destinations an application can write and other limitations, see Limits .

See: output smart constructor.

output

Arguments

:: Text
oName

-> DestinationSchema
oDestinationSchema

-> Output

Creates a value of Output with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

oKinesisStreamsOutput - Identifies an Amazon Kinesis stream as the destination.

oKinesisFirehoseOutput - Identifies an Amazon Kinesis Firehose delivery stream as the destination.

oName - Name of the in-application stream.

oDestinationSchema - Undocumented member.

oKinesisStreamsOutput :: Lens' Output (Maybe KinesisStreamsOutput)

Identifies an Amazon Kinesis stream as the destination.

oKinesisFirehoseOutput :: Lens' Output (Maybe KinesisFirehoseOutput)

Identifies an Amazon Kinesis Firehose delivery stream as the destination.

oName :: Lens' Output Text

Name of the in-application stream.

oDestinationSchema :: Lens' Output DestinationSchema

Undocumented member.

OutputDescription

data OutputDescription

Describes the application output configuration, which includes the in-application stream name and the destination where the stream data is written. The destination can be an Amazon Kinesis stream or an Amazon Kinesis Firehose delivery stream.

See: outputDescription smart constructor.

outputDescription :: OutputDescription

Creates a value of OutputDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

odOutputId - A unique identifier for the output configuration.

odDestinationSchema - Data format used for writing data to the destination.

odKinesisFirehoseOutputDescription - Describes the Amazon Kinesis Firehose delivery stream configured as the destination where output is written.

odKinesisStreamsOutputDescription - Describes Amazon Kinesis stream configured as the destination where output is written.

odName - Name of the in-application stream configured as output.

odOutputId :: Lens' OutputDescription (Maybe Text)

A unique identifier for the output configuration.

odDestinationSchema :: Lens' OutputDescription (Maybe DestinationSchema)

Data format used for writing data to the destination.

odKinesisFirehoseOutputDescription :: Lens' OutputDescription (Maybe KinesisFirehoseOutputDescription)

Describes the Amazon Kinesis Firehose delivery stream configured as the destination where output is written.

odKinesisStreamsOutputDescription :: Lens' OutputDescription (Maybe KinesisStreamsOutputDescription)

Describes Amazon Kinesis stream configured as the destination where output is written.

odName :: Lens' OutputDescription (Maybe Text)

Name of the in-application stream configured as output.

OutputUpdate

data OutputUpdate

Describes updates to the output configuration identified by the OutputId .

See: outputUpdate smart constructor.

outputUpdate

Arguments

:: Text
ouOutputId

-> OutputUpdate

Creates a value of OutputUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ouKinesisStreamsOutputUpdate - Describes an Amazon Kinesis stream as the destination for the output.

ouDestinationSchemaUpdate - Undocumented member.

ouKinesisFirehoseOutputUpdate - Describes a Amazon Kinesis Firehose delivery stream as the destination for the output.

ouNameUpdate - If you want to specify a different in-application stream for this output configuration, use this field to specify the new in-application stream name.

ouOutputId - Identifies the specific output configuration that you want to update.

ouKinesisStreamsOutputUpdate :: Lens' OutputUpdate (Maybe KinesisStreamsOutputUpdate)

Describes an Amazon Kinesis stream as the destination for the output.

ouDestinationSchemaUpdate :: Lens' OutputUpdate (Maybe DestinationSchema)

Undocumented member.

ouKinesisFirehoseOutputUpdate :: Lens' OutputUpdate (Maybe KinesisFirehoseOutputUpdate)

Describes a Amazon Kinesis Firehose delivery stream as the destination for the output.

ouNameUpdate :: Lens' OutputUpdate (Maybe Text)

If you want to specify a different in-application stream for this output configuration, use this field to specify the new in-application stream name.

ouOutputId :: Lens' OutputUpdate Text

Identifies the specific output configuration that you want to update.

RecordColumn

data RecordColumn

Describes the mapping of each data element in the streaming source to the corresponding column in the in-application stream.

Also used to describe the format of the reference data source.

See: recordColumn smart constructor.

recordColumn

Arguments

:: Text
rcName

-> Text
rcSqlType

-> RecordColumn

Creates a value of RecordColumn with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcMapping - Reference to the data element in the streaming input of the reference data source.

rcName - Name of the column created in the in-application input stream or reference table.

rcSqlType - Type of column created in the in-application input stream or reference table.

rcMapping :: Lens' RecordColumn (Maybe Text)

Reference to the data element in the streaming input of the reference data source.

rcName :: Lens' RecordColumn Text

Name of the column created in the in-application input stream or reference table.

rcSqlType :: Lens' RecordColumn Text

Type of column created in the in-application input stream or reference table.

RecordFormat

data RecordFormat

Describes the record format and relevant mapping information that should be applied to schematize the records on the stream.

See: recordFormat smart constructor.

recordFormat

Arguments

:: RecordFormatType
rfRecordFormatType

-> RecordFormat

Creates a value of RecordFormat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rfMappingParameters - Undocumented member.

rfRecordFormatType - The type of record format.

rfMappingParameters :: Lens' RecordFormat (Maybe MappingParameters)

Undocumented member.

rfRecordFormatType :: Lens' RecordFormat RecordFormatType

The type of record format.

ReferenceDataSource

data ReferenceDataSource

Describes the reference data source by providing the source information (S3 bucket name and object key name), the resulting in-application table name that is created, and the necessary schema to map the data elements in the Amazon S3 object to the in-application table.

See: referenceDataSource smart constructor.

referenceDataSource

Arguments

:: Text
rdsTableName

-> SourceSchema
rdsReferenceSchema

-> ReferenceDataSource

Creates a value of ReferenceDataSource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdsS3ReferenceDataSource - Undocumented member.

rdsTableName - Name of the in-application table to create.

rdsReferenceSchema - Undocumented member.

rdsS3ReferenceDataSource :: Lens' ReferenceDataSource (Maybe S3ReferenceDataSource)

Undocumented member.

rdsTableName :: Lens' ReferenceDataSource Text

Name of the in-application table to create.

rdsReferenceSchema :: Lens' ReferenceDataSource SourceSchema

Undocumented member.

ReferenceDataSourceDescription

data ReferenceDataSourceDescription

Describes the reference data source configured for an application.

See: referenceDataSourceDescription smart constructor.

referenceDataSourceDescription

Arguments

:: Text
rdsdReferenceId

-> Text
rdsdTableName

-> S3ReferenceDataSourceDescription
rdsdS3ReferenceDataSourceDescription

-> ReferenceDataSourceDescription

Creates a value of ReferenceDataSourceDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdsdReferenceSchema - Undocumented member.

rdsdReferenceId - ID of the reference data source. This is the ID that Amazon Kinesis Analytics assigns when you add the reference data source to your application using the AddApplicationReferenceDataSource operation.

rdsdTableName - The in-application table name created by the specific reference data source configuration.

rdsdS3ReferenceDataSourceDescription - Provides the S3 bucket name, the object key name that contains the reference data. It also provides the Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application reference table.

rdsdReferenceSchema :: Lens' ReferenceDataSourceDescription (Maybe SourceSchema)

Undocumented member.

rdsdReferenceId :: Lens' ReferenceDataSourceDescription Text

ID of the reference data source. This is the ID that Amazon Kinesis Analytics assigns when you add the reference data source to your application using the AddApplicationReferenceDataSource operation.

rdsdTableName :: Lens' ReferenceDataSourceDescription Text

The in-application table name created by the specific reference data source configuration.

rdsdS3ReferenceDataSourceDescription :: Lens' ReferenceDataSourceDescription S3ReferenceDataSourceDescription

Provides the S3 bucket name, the object key name that contains the reference data. It also provides the Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application reference table.

ReferenceDataSourceUpdate

data ReferenceDataSourceUpdate

When you update a reference data source configuration for an application, this object provides all the updated values (such as the source bucket name and object key name), the in-application table name that is created, and updated mapping information that maps the data in the Amazon S3 object to the in-application reference table that is created.

See: referenceDataSourceUpdate smart constructor.

referenceDataSourceUpdate

Arguments

:: Text
rdsuReferenceId

-> ReferenceDataSourceUpdate

Creates a value of ReferenceDataSourceUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdsuTableNameUpdate - In-application table name that is created by this update.

rdsuS3ReferenceDataSourceUpdate - Describes the S3 bucket name, object key name, and IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf and populate the in-application reference table.

rdsuReferenceSchemaUpdate - Undocumented member.

rdsuReferenceId - ID of the reference data source being updated. You can use the DescribeApplication operation to get this value.

rdsuTableNameUpdate :: Lens' ReferenceDataSourceUpdate (Maybe Text)

In-application table name that is created by this update.

rdsuS3ReferenceDataSourceUpdate :: Lens' ReferenceDataSourceUpdate (Maybe S3ReferenceDataSourceUpdate)

Describes the S3 bucket name, object key name, and IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf and populate the in-application reference table.

rdsuReferenceSchemaUpdate :: Lens' ReferenceDataSourceUpdate (Maybe SourceSchema)

Undocumented member.

rdsuReferenceId :: Lens' ReferenceDataSourceUpdate Text

ID of the reference data source being updated. You can use the DescribeApplication operation to get this value.

S3ReferenceDataSource

data S3ReferenceDataSource

Identifies the S3 bucket and object that contains the reference data. Also identifies the IAM role Amazon Kinesis Analytics can assume to read this object on your behalf.

An Amazon Kinesis Analytics application loads reference data only once. If the data changes, you call the UpdateApplication operation to trigger reloading of data into your application.

See: s3ReferenceDataSource smart constructor.

s3ReferenceDataSource

Arguments

:: Text
srdsBucketARN

-> Text
srdsFileKey

-> Text
srdsReferenceRoleARN

-> S3ReferenceDataSource

Creates a value of S3ReferenceDataSource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srdsBucketARN - Amazon Resource Name (ARN) of the S3 bucket.

srdsFileKey - Object key name containing reference data.

srdsReferenceRoleARN - ARN of the IAM role that the service can assume to read data on your behalf. This role must have permission for the s3:GetObject action on the object and trust policy that allows Amazon Kinesis Analytics service principal to assume this role.

srdsBucketARN :: Lens' S3ReferenceDataSource Text

Amazon Resource Name (ARN) of the S3 bucket.

srdsFileKey :: Lens' S3ReferenceDataSource Text

Object key name containing reference data.

srdsReferenceRoleARN :: Lens' S3ReferenceDataSource Text

ARN of the IAM role that the service can assume to read data on your behalf. This role must have permission for the s3:GetObject action on the object and trust policy that allows Amazon Kinesis Analytics service principal to assume this role.

S3ReferenceDataSourceDescription

data S3ReferenceDataSourceDescription

Provides the bucket name and object key name that stores the reference data.

See: s3ReferenceDataSourceDescription smart constructor.

s3ReferenceDataSourceDescription

Arguments

:: Text
srdsdBucketARN

-> Text
srdsdFileKey

-> Text
srdsdReferenceRoleARN

-> S3ReferenceDataSourceDescription

Creates a value of S3ReferenceDataSourceDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srdsdBucketARN - Amazon Resource Name (ARN) of the S3 bucket.

srdsdFileKey - Amazon S3 object key name.

srdsdReferenceRoleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf to populate the in-application reference table.

srdsdBucketARN :: Lens' S3ReferenceDataSourceDescription Text

Amazon Resource Name (ARN) of the S3 bucket.

srdsdFileKey :: Lens' S3ReferenceDataSourceDescription Text

Amazon S3 object key name.

srdsdReferenceRoleARN :: Lens' S3ReferenceDataSourceDescription Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf to populate the in-application reference table.

S3ReferenceDataSourceUpdate

data S3ReferenceDataSourceUpdate

Describes the S3 bucket name, object key name, and IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf and populate the in-application reference table.

See: s3ReferenceDataSourceUpdate smart constructor.

s3ReferenceDataSourceUpdate :: S3ReferenceDataSourceUpdate

Creates a value of S3ReferenceDataSourceUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srdsuBucketARNUpdate - Amazon Resource Name (ARN) of the S3 bucket.

srdsuFileKeyUpdate - Object key name.

srdsuReferenceRoleARNUpdate - ARN of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application.

srdsuBucketARNUpdate :: Lens' S3ReferenceDataSourceUpdate (Maybe Text)

Amazon Resource Name (ARN) of the S3 bucket.

srdsuFileKeyUpdate :: Lens' S3ReferenceDataSourceUpdate (Maybe Text)

Object key name.

srdsuReferenceRoleARNUpdate :: Lens' S3ReferenceDataSourceUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application.

SourceSchema

data SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

See: sourceSchema smart constructor.

sourceSchema

Arguments

:: RecordFormat
ssRecordFormat

-> NonEmpty RecordColumn
ssRecordColumns

-> SourceSchema

Creates a value of SourceSchema with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssRecordEncoding - Specifies the encoding of the records in the streaming source. For example, UTF-8.

ssRecordFormat - Specifies the format of the records on the streaming source.

ssRecordColumns - A list of RecordColumn objects.

ssRecordEncoding :: Lens' SourceSchema (Maybe Text)

Specifies the encoding of the records in the streaming source. For example, UTF-8.

ssRecordFormat :: Lens' SourceSchema RecordFormat

Specifies the format of the records on the streaming source.

ssRecordColumns :: Lens' SourceSchema (NonEmpty RecordColumn)

A list of RecordColumn objects.